pyproject.toml
[project]
name = "hswt"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "hydra-core>=1.3.0",
    "attn-gym>=0.0.4",
    "datasets>=2.0.0",
    "transformers>=4.30.0",
    "matplotlib>=3.9.4",
]

